refactor(docs): split up codeblock for readability
