+++ docs/src/pages/index.mdx
@@ -4,3 +4,23 @@ route: /
 ---
 
 # Hello World
+
+```typescript switcher
+---
+title: some title
+description: the description
+---
+const Comp = React.FC = () => <p>example 1</p>
+
+export default Comp
+
+---
+
+---
+title: some title 2
+description: the description
+---
+const Comp = React.FC = () => <p>example 2</p>
+
+export default Comp
+```
