+++ docs/tsconfig.json
@@ -1,12 +1,11 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "esModuleInterop": true,
     "jsx": "react",
     "outDir": "dist",
     "plugins": [{ "name": "typescript-plugin-css-modules" }],
     "rootDir": "src",
-    "target": "es5"
+    "skipLibCheck": true
   },
   "exclude": ["*.js", "dist", "public", "node_modules"]
 }
