+++ tsconfig.json
@@ -9,6 +9,7 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "outDir": "dist",
+    "resolveJsonModule": true,
     "rootDir": "src",
     "skipLibCheck": true,
     "sourceMap": true,
